@@ -1,6 +1,4 @@
-﻿using Shared.Events;
-
-namespace Eiromplays.IdentityServer.Application.Common.Events;
+﻿namespace Eiromplays.IdentityServer.Application.Common.Events;
 
 // This is just a shorthand to make it a bit easier to create event handlers for specific events.
 public interface IEventNotificationHandler<TEvent> : INotificationHandler<EventNotification<TEvent>>
@@ -12,7 +10,7 @@ public abstract class EventNotificationHandler<TEvent> : INotificationHandler<Ev
     where TEvent : IEvent
 {
     public Task Handle(EventNotification<TEvent> notification, CancellationToken cancellationToken) =>
-        Handle(notification.Event, cancellationToken);
+        HandleAsync(notification.Event, cancellationToken);
 
-    public abstract Task Handle(TEvent @event, CancellationToken cancellationToken);
+    public abstract Task HandleAsync(TEvent @event, CancellationToken cancellationToken);
 }
