@@ -2,9 +2,9 @@ namespace Eiromplays.IdentityServer.Application.Identity.Users.Password;
 
 public class ResetPasswordRequest
 {
-    public string? Email { get; set; }
+    public string Email { get; set; } = default!;
 
-    public string? Password { get; set; }
+    public string Password { get; set; } = default!;
 
-    public string? Token { get; set; }
+    public string Token { get; set; } = default!;
 }
