@@ -1,11 +1,11 @@
-using PhoneNumbers;
-
 namespace Eiromplays.IdentityServer.Application.Identity.Users;
 
 public class UpdateUserRequestValidator : Validator<UpdateUserRequest>
 {
-    public UpdateUserRequestValidator(IUserService userService, IStringLocalizer<UpdateUserRequestValidator> T, IValidator<FileUploadRequest?> fileUploadRequestValidator)
+    public UpdateUserRequestValidator(IStringLocalizer<UpdateUserRequestValidator> T)
     {
+        var fileUploadRequestValidator = Resolve<IValidator<FileUploadRequest?>>();
+
         RuleFor(p => p.Id)
             .NotEmpty();
 
@@ -21,7 +21,11 @@ public class UpdateUserRequestValidator : Validator<UpdateUserRequest>
             .NotEmpty()
             .EmailAddress()
                 .WithMessage(T["Invalid Email Address."])
-            .MustAsync(async (user, email, _) => !await userService.ExistsWithEmailAsync(email, user.Id))
+            .MustAsync(async (user, email, _) =>
+            {
+                var userService = Resolve<IUserService>();
+                return !await userService.ExistsWithEmailAsync(email, user.Id);
+            })
                 .WithMessage((_, email) => string.Format(T["Email {0} is already registered."], email));
 
         RuleFor(p => p.Image)
@@ -29,7 +33,11 @@ public class UpdateUserRequestValidator : Validator<UpdateUserRequest>
 
         var phoneNumberUtil = PhoneNumbers.PhoneNumberUtil.GetInstance();
         RuleFor(u => u.PhoneNumber).Cascade(CascadeMode.Stop)
-            .MustAsync(async (user, phone, _) => !await userService.ExistsWithPhoneNumberAsync(phone!, user.Id))
+            .MustAsync(async (user, phone, _) =>
+            {
+                var userService = Resolve<IUserService>();
+                return !await userService.ExistsWithPhoneNumberAsync(phone!, user.Id);
+            })
                 .WithMessage((_, phone) => string.Format(T["Phone number {0} is already registered."], phone))
             .Must(phoneNumber => phoneNumberUtil.IsValidNumber(phoneNumberUtil.Parse(phoneNumber, null)))
                 .WithMessage(T["Invalid Phone Number."])
