@@ -100,12 +100,7 @@ export default defineConfig({
         changeOrigin: false,
         secure: false,
       },
-      '/products': {
-        target: target,
-        changeOrigin: false,
-        secure: false,
-      },
-      '/brands': {
+      '/identity-providers': {
         target: target,
         changeOrigin: false,
         secure: false,
