@@ -1,3 +1,4 @@
+using System.Text.Json;
 using Eiromplays.IdentityServer.Application.Identity.Auth;
 using Eiromplays.IdentityServer.Application.Identity.Auth.Requests.ExternalLogins;
 using Microsoft.AspNetCore.Authentication;
@@ -7,10 +8,12 @@ namespace Eiromplays.IdentityServer.Endpoints.v1.Account;
 public class ExternalLoginEndpoint : Endpoint<ExternalLoginRequest, AuthenticationProperties>
 {
     private readonly IAuthService _authService;
+    private readonly ILogger<ExternalLoginEndpoint> _logger;
 
-    public ExternalLoginEndpoint(IAuthService authService)
+    public ExternalLoginEndpoint(IAuthService authService, ILogger<ExternalLoginEndpoint> logger)
     {
         _authService = authService;
+        _logger = logger;
     }
 
     public override void Configure()
@@ -27,7 +30,15 @@ public class ExternalLoginEndpoint : Endpoint<ExternalLoginRequest, Authenticati
 
     public override async Task HandleAsync(ExternalLoginRequest req, CancellationToken ct)
     {
+        var authenticationProperties = await _authService.ExternalLoginAsync<ExternalLoginCallbackEndpoint>(req, HttpContext.Response);
+
+        if (HttpContext.Response.HasStarted)
+        {
+            _logger.LogWarning("The response has already started, the external login middleware will not be able to redirect the browser");
+            return;
+        }
+
         await SendOkAsync(
-            await _authService.ExternalLoginAsync<ExternalLoginCallbackEndpoint>(req, HttpContext.Response), ct);
+            authenticationProperties, ct);
     }
 }
